Policy

+++ app/Http/Controllers/UserController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Middleware\OnceBasicMiddleware;
 use App\Http\Resources\UserResource;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -13,7 +12,7 @@ class UserController extends Controller
 
     public function __construct()
     {
-        $this->middleware("auth:sanctum")->except("index", "store");
+        $this->middleware("auth:sanctum")->except("index");
     }
     /**
      * Display a listing of the resource.
@@ -29,6 +28,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize("create", User::class);
          $data = $request->all();
          $data["password"] = Hash::make($data["password"]);
          $user = new UserResource(User::create($data));
@@ -49,6 +49,7 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize("update", User::findOrFail($id));
         $user = new UserResource(User::findOrFail($id));
         $user->update($request->all());
         return  $user->response()->setStatusCode(200, "Updated Successffully");
@@ -59,6 +60,7 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize("delete", User::findOrFail($id));
         User::findOrFail($id)->delete();
         return response()->json([
             "data" => "deleted successfully",
